pyproject.toml
[project]
name = "aria"
version = "0.1.0"
description = "Autonomous Research & Intelligence Analyst — M&A due diligence multi-agent system"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "openai>=1.0.0",
    "yfinance>=0.2.40",
    "httpx>=0.27.0",
    "tavily-python>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
postgres = [
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
